Fix #81: BigInt serialization in DebugSnapshot
Changed StatsInfo.ticks from u64 to u32 to prevent BigInt
in JS that JSON.stringify cannot serialize. u64 in wasm-bindgen
structs becomes JS BigInt which breaks DebugSnapshot.capture()
at line 7358 (JSON.stringify(info)).

Cache busted from v=74 to v=75.

+++ engine/src/lib.rs
@@ -3430,7 +3430,7 @@ pub fn on_wheel(delta_y: f32) {
 #[derive(Copy, Clone)]
 pub struct StatsInfo {
     pub fps: u32,
-    pub ticks: u64,
+    pub ticks: u32,
     pub game_time: f32,
     pub zoom: f32,
 }
@@ -3441,7 +3441,7 @@ pub fn get_stats() -> Option<StatsInfo> {
     unsafe {
         (*std::ptr::addr_of!(APP)).as_ref().map(|app| StatsInfo {
             fps: app.current_fps,
-            ticks: app.game_loop.state.tick_count,
+            ticks: app.game_loop.state.tick_count as u32,
             game_time: app.game_loop.state.game_time as f32,
             zoom: app.camera.zoom,
         })
